refactor : move security excp
