+++ backend/src/main/java/toonpick/app/exception/exception/InvalidJwtTokenException.java
@@ -1,4 +1,4 @@
-package toonpick.app.security.exception;
+package toonpick.app.exception.exception;
 
 public class InvalidJwtTokenException extends JwtException {
     public InvalidJwtTokenException(String message) {
