+++ backend/src/main/java/toonpick/app/security/controller/TokenReissueController.java
@@ -11,9 +11,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import toonpick.app.security.exception.ExpiredJwtTokenException;
-import toonpick.app.security.exception.InvalidJwtTokenException;
-import toonpick.app.security.exception.MissingJwtTokenException;
+import toonpick.app.exception.exception.ExpiredJwtTokenException;
+import toonpick.app.exception.exception.InvalidJwtTokenException;
+import toonpick.app.exception.exception.MissingJwtTokenException;
 import toonpick.app.security.jwt.JwtTokenProvider;
 import toonpick.app.security.jwt.JwtTokenValidator;
 import toonpick.app.security.token.TokenService;
